refactor: cleanup and move digest impl to separate file
