+++ ethers-core/src/types/crypto/mod.rs
@@ -3,3 +3,5 @@ pub use keys::{PrivateKey, PublicKey};
 
 mod signature;
 pub use signature::Signature;
+
+mod hash;
